# Add repeat for Char to fix performance of pow