--- a/doc/src/stdlib/linalg.md
+++ b/doc/src/stdlib/linalg.md
@@ -89,7 +89,7 @@ Base.inv(::AbstractMatrix)
 Base.LinAlg.pinv
 Base.LinAlg.nullspace
 Base.repmat
-Base.repeat
+Base.repeat(::AbstractArray)
 Base.kron
 Base.SparseArrays.blkdiag
 Base.LinAlg.linreg