close https://github.com/JuliaLang/julia/issues/22785

But more importantly fix the bug for  `::Char ^ 0` and `::Char^1`

As a future optimization someone may want to improve the nonascii case to remove the allocation for the string conversion. 


```julia
# benchmark using 0.6

julia> using BenchmarkTools; import Base.repeat, Base._string_n

julia> function repeat(c::Char, r::Integer)
           r < 0 && throw(ArgumentError("can't repeat a character $r times"))
           if isascii(c)
               out = _string_n(r)
               ccall(:memset, Ptr{Void}, (Ptr{UInt8}, Cint, Csize_t), out, c, r)
               return out
           else
               return repeat(string(c), r)
           end
       end
repeat (generic function with 4 methods)

julia> @btime repeat($('x'), $(100));
  28.140 ns (1 allocation: 128 bytes)

julia> @btime repeat($("x"), $(100));
  280.596 ns (1 allocation: 128 bytes)

julia> @btime repeat($('x'), $(10));
  18.158 ns (1 allocation: 32 bytes)

julia> @btime repeat($("x"), $(10));
  42.239 ns (1 allocation: 32 bytes)

julia> @btime repeat($('x'), $(2));
  18.158 ns (1 allocation: 32 bytes)

julia> @btime repeat($("x"), $(2));
  21.711 ns (1 allocation: 32 bytes)
```