@musm 2017-07-13
Interestingly the performance characteristics of `repeat` on master have changed substantially than from `0.6`.