@musm 2017-07-14
Updated to also test  `AbstractString`  subtypes