@musm 2017-07-14
@tkelman and @stevengj  I have addressed your reviews

(please, don't forget to squash merge)