@stevengj 2017-07-15
(Travis failure is #17626.)