@musm 2017-07-15
I think there is a bug in the implentation here
e.g. try this on `$` , which spits

`"\$\$\$\$\$\$\$\$\$\$\$\$\$\$\$\$\$\$\$\$"`