@TotalVerb 2017-07-15
@musm That's because `$` must be escaped within double quotes.