@musm 2017-07-15
@TotalVerb  oh right, thanks!

I added a new commit which removed the `GenericString` test by accident, I dropped the commit.