@KristofferC 2017-07-16
Good to go then?